GltfMeshUtility を基底クラスとして切り出し
