+++ Assets/UniGLTF/Runtime/MeshUtility/GltfMeshUtility.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: e2425cf6ac1f2434986968ff0d4a4755
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
