+++ Assets/UniGLTF/Runtime/MeshUtility/MeshIntegrationGroup.cs
@@ -1,5 +1,5 @@
+using System;
 using System.Collections.Generic;
-using System.Linq;
 using UnityEngine;
 
 namespace UniGLTF.MeshUtility
@@ -8,5 +8,10 @@ namespace UniGLTF.MeshUtility
     {
         public string Name;
         public List<Renderer> Renderers = new List<Renderer>();
+
+        public static List<MeshIntegrationGroup> ToList()
+        {
+            throw new NotImplementedException();
+        }
     }
 }
